pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cloud-api-sample-code"
version = "0.1.0"
description = "Firstbeat Sports Cloud API - Sample Code and Reference Implementations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "numpy>=2.4.1",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
]

[tool.hatch.build.targets.wheel]
packages = ["src/python_sync_app"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[dependency-groups]
dev = [
    "coverage>=7.13.1",
    "openapi-core>=0.22.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
